fix: runner crash detection + remaining dtype bugs for bf16
Runner: add pipefail to tee pipeline, ghost crash handler (exit 0 +
no .json), corrupt JSON handler, normalize .failed file format.

Dtype: SLOT delta/logit_bias match model dtype, SSM init_state
matches embed dtype, position arange uses float32 not bf16.

+++ src/chaoscontrol/memory.py
@@ -389,7 +389,7 @@ class MultiSlotOuterModel(nn.Module):
         """
         batch, seq, dim = h_seq.shape
         # Exponential recency weights: later positions matter more
-        positions = torch.arange(seq, dtype=h_seq.dtype, device=h_seq.device)
+        positions = torch.arange(seq, dtype=torch.float32, device=h_seq.device)
         weights = torch.exp(positions - positions[-1])  # exp-decay, last position = weight 1.0
         weights = weights / weights.sum()
         # Weighted sum over sequence: (batch, dim)
